feat: attempt to make proper router and slug, with errors now
