--- search-app/src/pages/[slug]/Details.tsx
@@ -1,8 +1,0 @@
-import { SideSection } from "@/components/main/sideSection/sideSection";
-import { IPerson } from "@/types/types";
-
-export default function Details({ person }: { person: IPerson }) {
-  const slug = person.name;
-
-  return <SideSection params={{ slug }} />;
-}
