--- search-app/src/pages/[slug]/page.tsx
@@ -1,10 +1,0 @@
-import "../../index.css";
-import { ClientOnly } from "./client";
-
-export function generateStaticParams() {
-  return [{ slug: [""] }];
-}
-
-export default function Page() {
-  return <ClientOnly />;
-}
