+++ search-app/src/pages/layout.tsx
@@ -1,22 +1,5 @@
-import "./index.css";
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
+import LayoutComponent from "@/components/layoutComponent/layoutComponent";
 
-const inter = Inter({ subsets: ["latin"] });
-
-export const metadata: Metadata = {
-  title: "Search Star Wars People",
-  description: "Web site created with Next.js.",
-};
-
-export default function RootLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <html lang="en">
-      <body>
-        <div id="root" className={inter.className}>
-          {children}
-        </div>
-      </body>
-    </html>
-  );
+export default function Layout({ children }: { children: React.ReactNode }) {
+  return <LayoutComponent>{children}</LayoutComponent>;
 }
